refactor(CvView): moving ActionButton to its own file
